Add files via upload
In this project I want to check the following in udemy website,
I collected the data from kaggle and it is a csv file. using 
pandas I read in the csv into my jupiter notebook with 
pd.read_csv. i proceeded to analysing the basic information I 
needed from the dataset.
Using the head and tail function in python I checked the top5 
and bottom5 of my dataset

1. What are the different subjects for which udemy is offering 
courses? There are just 4 subject in the dataset, (Which are 
Musical Instruments, Business Finance, Graphic Design, Web 
Development)

2. What subject has the maximum courses. The output shows that.. 
Web Development   has     1200
Business Finance   has    1199
Musical Instruments  has   680
Graphic Design      has    603

3. Show the all the free courses. From the output the free courses are 310
4. Show all the paid courses. From the output the free courses are 3372
5. Which are the top selling courses. In findidng the top selling
courses I used the number of subscribers as the yardstick, I 
believe the higher the subscribers the higher the higher the 
purchase = sales. Web Development is TOP 5 in "Number of Subscribers"

6. Which are the least selling courses? Business Finance and 
Graphics Design made it the BOTTOM 5 in "Number of Subscribers"
7. Show all course the Web Development/ Business Finance where 
the price is equal to or greater 100.
8. show all courses List out all the courses that are related to
python/sql
9. Show the counts of number of courses containing Python/SQL 
10. What are courses published in the year 2015/2020/2023
11. What are the max nmber of subscribers for each level of 
courses.